CropDisease prediction/backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
python-multipart
requests
deep-translator
